Refactor session management and improve product handling in admin pages; streamline input validation and enhance CSS styling for better layout.

+++ admin/delete.php
@@ -1,11 +1,18 @@
 <?php
+session_start();
+if (!isset($_SESSION['login'])) {
+  header("Location: login.php");
+  exit();
+}
+?>
+<?php
 include 'config.php';
 
 $id = $_GET['id'];
 $sql = "DELETE FROM produk WHERE id=$id";
 
 if ($conn->query($sql) === TRUE) {
-  header("Location: index.php"); // kembali setelah hapus
+  header("Location: adminindex.php"); // kembali setelah hapus
 } else {
   echo "Error: " . $sql . "<br>" . $conn->error;
 }
